requirements-ci.txt
pyserial==3.5
pytest==7.4.3
pytest-cov==4.1.0
numpy>=1.21.0
# PyQt5 and pyqtgraph are mocked in CI/CD environments (conftest.py)
